Added clipboard based uploading
When an image is stored in clipboard, this is detected, and it can be
uploaded.

Similarly, when one or more files are copied to the clipboard, their
contents are checked if they are an image. If so, they can all be
uploaded.

This contains a conversion to VS 2015, and some cleanups.